feat: rename hql-span to hql-core and demote span to module
