+++ Cargo.toml
@@ -85,9 +85,9 @@ hash-repo-chores.path               = "libs/@local/repo-chores/rust"
 hash-status.path                    = "libs/@local/status/rust"
 hash-temporal-client.path           = "libs/@local/temporal-client"
 hash-tracing.path                   = "libs/@local/tracing"
+hql-core.path                       = "libs/@local/hql/core"
 hql-cst.path                        = "libs/@local/hql/cst"
 hql-diagnostics.path                = "libs/@local/hql/diagnostics"
-hql-span.path                       = "libs/@local/hql/span"
 hql-symbol.path                     = "libs/@local/hql/symbol"
 type-system.path                    = "libs/@blockprotocol/type-system/rust"
