+++ libs/@local/hql/core/LICENSE.md
