+++ libs/@local/hql/core/docs/dependency-diagram.mmd
