+++ libs/@local/hql/core/package.json
