+++ libs/@local/hql/core/src/lib.rs
@@ -1,0 +1,9 @@
+//! # HashQL Core
+//!
+//! ## Workspace dependencies
+#![cfg_attr(doc, doc = simple_mermaid::mermaid!("../docs/dependency-diagram.mmd"))]
+#![feature(never_type)]
+
+extern crate alloc;
+
+pub mod span;
