+++ libs/@local/hql/core/src/span/entry.rs
