+++ libs/@local/hql/core/src/span/node.rs
