+++ libs/@local/hql/symbol/Cargo.toml
@@ -8,7 +8,7 @@ version.workspace = true
 
 [dependencies]
 # Public workspace dependencies
-hql-span = { workspace = true, public = true }
+hql-core = { workspace = true, public = true }
 
 # Public third-party dependencies
