+++ libs/@local/hql/syntax-jexpr/Cargo.toml
@@ -8,9 +8,9 @@ publish.workspace = true
 
 [dependencies]
 # Public workspace dependencies
+hql-core        = { workspace = true, public = true }
 hql-cst         = { workspace = true, public = true }
 hql-diagnostics = { workspace = true, public = true }
-hql-span        = { workspace = true, public = true }
 
 # Public third-party dependencies
 jsonptr   = { workspace = true, public = true }
